💄 Updated inputs' story
